mfc(MFC-AUDIT): ask the binary which MFC container it is (CRuntimeClass) + fix 96 wrong-class bindings
FID cannot name MFC's containers: CObArray/CPtrArray/CDWordArray/CByteArray are
byte-identical, CObList/CPtrList/CStringList are byte-identical, the map families
are byte-identical.  There is NO COMDAT fold (MSVC5 has no /OPT:ICF) - the linker
emits every one of them, they just look the same - so a byte-signature matcher can
only ever say AMBIG, and the tree had been trusting those rows and getting them
wrong.  objdiff masks relocations, so a wrong class costs ~0% and displays 100.00%
while being a link defect and a silent wrong-symbol binding at runtime.

The binary names them itself.  IMPLEMENT_DYNAMIC/SERIAL emits a CRuntimeClass whose
FIRST field is the class-name string, and CObject's first virtual is GetRuntimeClass
- so vtable slot 0 is a 6-byte `mov eax,&classCXxx; ret`.  Follow it and the class
states its own name.  Each container's ctor stamps its OWN vtable, so each .obj band
is self-identifying.

THE TOOL   scripts/gruntz/analysis/mfc_class.py
  python -m gruntz.analysis.mfc_class <rva>        # who owns this? + evidence chain
  python -m gruntz.analysis.mfc_class --map        # 65 CRuntimeClasses, 74 vtables, bands
  python -m gruntz.analysis.mfc_class --audit      # audit the TREE against retail
  python -m gruntz.analysis.mfc_class --relabel    # fix the AMBIG FID rows (wired into
                                                   # fid_generate as stage 6)
--audit pairs our base COMDAT's reloc records against retail's bytes at the same
offset: "we emit ?Lookup@CMapStringToOb; does the address retail calls live in the
CMapStringToOb band?".  A NO is a binding objdiff cannot see.

RECOVERED BANDS (each ctor DIR32s its own vtable -> it names itself)
  CPtrList  0x1b4867/0x1eb054   CObList     0x1b59cc/0x1ed4b4   CStringList 0x1b5d04/0x1ed4dc
  CDWordArr 0x1b4b43/0x1ec29c   CPtrArray   0x1b4f0b/0x1ec2dc
  CByteArr  0x1b527e/0x1ed28c   CObArray    0x1b55e9/0x1ed494
  CMapStringToOb 0x1b7e17/0x1eafd4 (Lookup 0x1b8008)
  CMapStringToPtr 0x1b8247/0x1eb014 (Lookup 0x1b8438)   CMapPtrToPtr 0x1b85b1/0x1ed264

FIXED (all binary-proven, per site - never blanket)
  * 39 CObList -> CPtrList: CSpawnList, GzObList/CGruntzCmdList, CKeyedList/CNetThing,
    CImagePool's two lists, CTriggerMgr's 12, CTileTriggerContainer, CWorldSoundSet,
    CNetCmdSlot::m_cmds.  Their (CObject*) AddTail casts were forced by the WRONG
    container and are gone - the payloads are not CObject-derived.
  * CMultiStartDlg::m_74 is a ::CStringList (ctor 0x1b5d04), not CObList - the exact
    "ctor calls band C, dtor calls the CStringList dtor" silent-misbind shape.
  * 38 CMapStringToOb <-> CMapStringToPtr, BOTH directions: the tree believed
    0x1b8008/0x1b8438 were one COMDAT-folded body.  They are two classes.  0x1b8008 IS
    CMapStringToOb::Lookup; 0x1b8438 IS CMapStringToPtr::Lookup.  Fixed the canonical
    CSndHost::m_10 (~60 TUs), CDDrawSubMgrLeaf/LeafScan, CMenuItem, C9cab0, ChatBox(+Owner).
    CDDrawWorkerRegistry::m_map really IS CMapStringToOb - left alone.
  * CGameLevel's three arrays are ::CObArray (ctor 0x15ccd0 calls 0x1b55e9 x3).  This
    OVERTURNS the documented "genuine CDWordArray" reading: the old CTypedPtrArray<
    CPtrArray> probe only proved it was not CPtrArray.
  * DinMgr2's `(CDWordArray*)` cast on a CPtrArray - adopted *because* FID said
    CDWordArray - is gone.  Same for CGruntzMapMgr::m_arr, CGruntSpawnConfig::m_18.
  * views dissolved onto real MFC classes: CMapPtrArray -> ::CPtrArray,
    CMovieByteArray -> ::CDWordArray.

ROOT CAUSE FIXED: config/library_labels.csv - 98 of 186 NAFXCW rows in the container
bands named the wrong class.  Now 176/176 correct (75 re-classed, 5 fid-noise dtors
re-spelled, 10 FABRICATED rows dropped - ?SetSize@CStateStackZ / ?AddTail@GzObList,
symbols NAFXCW does not define, hand-added to paper over fake views).  fid_generate
now runs --relabel as stage 6 so a regeneration cannot re-introduce them.

NUMBERS (one environment, fresh builds both sides)
  mfc_class --audit  offset-paired wrong-class sites : 111 -> 15
                     class-SET disagreements         : 241 -> 106
  library_labels container rows correct              :  88/186 -> 176/176
  reloc_fidelity     7 MISBOUND / 77 UNBOUND (unchanged - these are DATA/fake-decl
                     defects, a different axis; the MFC misbindings were all CALL
                     targets to *real* library rvas, which reloc_fidelity marks EXEMPT)
  link_defects       586 (unchanged: 302 PHANTOM / 284 UNDEFINED-DATA / 0 MULTIPLY / 15 DIVERGENT)
  objdiff            2304 -> 2303 exact, 72.01% -> 72.00% fuzzy.  MAX % unchanged (100.00%).
                     The -1: binding to the CORRECT NAFXCW symbol re-pairs the delink
                     and flipped one function out of exact.  The lie scored better than
                     the truth; truth kept.
  gates              class-size OK, vtbl-uniqueness OK, vtable-owner 0 MISBOUND, rc=0
  )this casts 382 -> 381.  )m_ casts unchanged at 1201 - and that is AUTHENTIC:
  CPtrList/CMapStringToPtr are genuine void* containers, so a typed element really does
  need a cast.  The CObList/CObject* version was the lie.

STILL WRONG (15 sites, honest gap - run `mfc_class --audit` for the list): gruntzmgr
CDemo/CPlay (CByteArray->CPtrArray) + CheatRevealTreasures, wwdgameobject x4,
bootystateactivate, gruntcombat, ingameicon, movieplayer, worldsoundset.

(cherry picked from commit 471da3dc938df25aa6d7321b39ec1c233c297633)

+++ src/Gruntz/GruntSpawnConfig.cpp
@@ -444,7 +444,7 @@ BOOL CGruntSpawnConfig::BuildVoiceList() {
     m_18.SetSize(0, -1);
     m_18.SetAtGrow(0, 0);
     for (i32 i = 1; i < 0x4b0; i++) {
-        m_18.SetAtGrow(i, (DWORD)BuildVoiceSoundList(i));
+        m_18.SetAtGrow(i, BuildVoiceSoundList(i));
     }
     return 1;
 }
@@ -457,8 +457,9 @@ BOOL CGruntSpawnConfig::BuildVoiceList() {
 // and append it to the list (AddTail). The by-value CString param + the inner
 // copy construct the /GX frame. EXACT once the flag is passed through (the
 // earlier "frame-size wall" was really the dropped 2nd ctor arg - a `push flag`).
-// The (CObject*) cast is language-forced: CPtrList stores raw non-CObject records
-// here exactly as retail does.
+// The list is a CPtrList (retail calls 0x1b4867/0x1b4991, the CPtrList band - see
+// SpawnList.h), so AddTail takes the record straight: the old `(CObject*)` cast was
+// forced by the WRONG container type, not by the language.
 RVA(0x0011c560, 0x91)
 void CSpawnList::AddVoiceSound(CString s, i32 flag) {
     CSpawnEntry* node = new CSpawnEntry(s, flag);
